Move EmojiButton to its own component file.
