--- packages/editor/src/client/blocks/emoji-mart-override.css
@@ -1,12 +1,0 @@
- .emoji-mart-bar:last-child {
-   display: none;
- }
-
- .emoji-mart-scroll {
-   overflow-y: scroll;
-   overflow-x: hidden;
-   /* override height */
-   height: 12em;
-   padding: 0 6px 6px 6px;
-   will-change: transform;
- }
